Started to change ServiceQueue to handle Promises and Reakt callbacks.
